// src/types/invidious.ts
export interface ThumbnailObject {
  quality: string;
  url: string;
  width: number;
  height: number;
}
// ...
export interface VideoDetails {
  type: string;
  title: string;
  videoId: string;
  videoThumbnails: ThumbnailObject[];
  storyboards?: { url: string; templateUrl: string; width: number; height: number; count: number; interval: number; }[];
  description: string;
  descriptionHtml: string;
  published: number;
  publishedText: string;
  keywords: string[];
  viewCount: number;
  likeCount: number;
  author: string;
  authorId: string;
  authorUrl: string;
  authorThumbnails: ThumbnailObject[];
  subCountText: string;
  lengthSeconds: number;
  allowRatings: boolean;
  rating: number;
  isListed: boolean;
  liveNow: boolean;
  isUpcoming: boolean;
  premiereTimestamp?: number;
  hlsUrl?: string;
  dashUrl?: string;
  formatStreams: FormatStream[];
  adaptiveFormats: AdaptiveFormat[];
  captions: Caption[];
  recommendedVideos: RecommendedVideo[];
}

export interface FormatStream {
  url: string;
  mimeType: string;
  container: string;
  encoding: string;
  quality: string;
  qualityLabel: string;
  resolution: string;
  fps: number;
  bitrate: string;
  clen?: string;
}

export interface AdaptiveFormat {
  index: string;
  bitrate: string;
  init: string;
  url: string;
  mimeType: string;
  container: string;
  encoding: string;
  qualityLabel: string;
  resolution: string;
  fps?: number;
  audioQuality?: string;
  audioSampleRate?: number;
  audioChannels?: number;
}

export interface Caption {
  label: string;
  languageCode: string;
  url: string;
}

export interface RecommendedVideo {
  videoId: string;
  title: string;
  author: string;
  authorId: string;
  authorThumbnails: ThumbnailObject[];
  videoThumbnails: ThumbnailObject[];
  viewCount: number;
  lengthSeconds: number;
}